pyproject.toml
[tool.poetry]
name = "ai-assistant"
version = "0.1.0"
description = ""
authors = ["Hafsa"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
streamlit = "^1.38.0"
longchain = "^0.4.1"
langchain-community = "^0.2.15"
python-dotenv = "^1.0.1"
langchain-experimental = "^0.0.64"
openai = "^1.43.0"
wikipedia = "^1.4.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
